CPU Play
I added all the sub-board combinations, but until we decide on a hash function for the sub-boards, the negamax is just evaluating each sub-board board when it gets to it. Also, I made some changes to the play method so you can play vs the negamax. Currently, the depth count is capped at 3, but we can change that to be limited by time instead.